feat(ffi): Extract `TaskHandle` into itw own module.
